refactor: create domain package
