+++ src/main/java/race/domain/MoveCars.java
@@ -1,4 +1,4 @@
-package race;
+package race.domain;
 
 import java.util.Collection;
