+++ src/main/java/race/view/outbound/OutputView.java
@@ -1,6 +1,6 @@
 package race.view.outbound;
 
-import race.Car;
+import race.domain.Car;
 
 import java.util.Collection;
 import java.util.stream.Collectors;
